Added a feature to choose for only upload image if a "Image Link" column was found in the file to import

+++ assets/js/components/views/Import.vue
@@ -159,22 +159,36 @@
             </v-stepper-content>
 
             <v-stepper-content step="5" class="text-center">
-                <div v-if="data.cleaningUp">
-                    <v-progress-circular
-                        :value="((data.unimported.deleted / data.unimported.items.length) * 100)" 
-                        :size="150"
-                        :width="20"
-                        color="orange"></v-progress-circular>
-                    <h3 class="mb-0" v-if="data.cleaningUp">Cleaning up...</h3>
+                <div v-if="data.hasImageLinks && !data.importing">
+                    <h3>Your file contains column for image links. Are you going to upload images only and not import products?</h3>
+                    <v-btn
+                        @click="data.importProducts=false;doImport(5, true)"  
+                        color="primary" 
+                        class="mb-3 mr-1">Yes</v-btn>
+                    <v-btn 
+                        @click="data.importProducts=true;doImport(5, true)" 
+                        color="error" 
+                        class="mb-3 ml-1">No</v-btn>
                 </div>
 
-                <div v-else>
-                    <v-progress-circular
-                        :value="((data.imported.items.length / data.items.length) * 100)" 
-                        :size="150"
-                        :width="20"
-                        color="green"></v-progress-circular>
-                    <h3 class="mb-0">{{ data.imported.items.length }} of {{ data.items.length }}</h3>
+                <div v-if="data.importing">
+                    <div v-if="data.cleaningUp">
+                        <v-progress-circular
+                            :value="((data.unimported.deleted / data.unimported.items.length) * 100)" 
+                            :size="150"
+                            :width="20"
+                            color="orange"></v-progress-circular>
+                        <h3 class="mb-0" v-if="data.cleaningUp">Cleaning up...</h3>
+                    </div>
+
+                    <div v-else>
+                        <v-progress-circular
+                            :value="((data.imported.items.length / data.items.length) * 100)" 
+                            :size="150"
+                            :width="20"
+                            color="green"></v-progress-circular>
+                        <h3 class="mb-0">{{ data.imported.items.length }} of {{ data.items.length }}</h3>
+                    </div>
                 </div>
             </v-stepper-content>
 
@@ -187,9 +201,12 @@
                     </v-list-item-avatar>
                     <v-list-item-content>
                         <v-list-item-title>{{ item.title }}</v-list-item-title>
-                        <v-list-item-subtitle>Product ID: #{{ item.id }} | SKU: {{ item.sku }}</v-list-item-subtitle>
+                        <v-list-item-subtitle 
+                            v-if="!item.is_image">Product ID: #{{ item.id }} | SKU: {{ item.sku }}</v-list-item-subtitle>
                     </v-list-item-content>
-                    <v-list-item-action class="mx-0 d-block">
+                    <v-list-item-action 
+                        v-if="!item.is_image" 
+                        class="mx-0 d-block">
                         <v-btn 
                             v-if="item.url" 
                             :href="item.url" 
@@ -248,7 +265,10 @@
                 data    : {
                     time            : new Date().getTime(),
                     loading         : false,
+                    importing       : false,
                     cleaningUp      : false,
+                    hasImageLinks   : false,
+                    importProducts  : true,
                     imagesSource    : [],
                     items           : [],
                     sources         : [],
@@ -268,6 +288,7 @@
                             {key:'sale_price', selected:null},
                             {key:'tax_class', selected:'GST'},
                             {key:'supplier_part_number', selected:'Supplier Part No.'},
+                            {key:'image_link', selected:'Image Link'},
                             {key:'barcode', selected:'Barcode'},
                             {key:'inner_barcode', selected:'Inner Barcode'},
                             {key:'uom', selected:'Units'},
@@ -370,6 +391,11 @@
                                     instance.data.map.headers.push(h == 0 ? '__EMPTY' : '__EMPTY_'+ h)
                                 }
                             }
+
+                            if( 0 <= instance.data.map.headers.indexOf('Image Link') ) {
+                                instance.data.hasImageLinks = true
+                            }
+
                             instance.data.items = XLSX.utils.sheet_to_json(workbook.Sheets[sheetName])
                         })
                     }
@@ -382,7 +408,12 @@
                         instance.data.loading = false
 
                         if( instance.models.skipMapping && instance.models.skipImageSource ) {
-                            instance.doImport(5)
+                            if( !instance.data.hasImageLinks ) {
+                                instance.doImport(5, true)
+                                instance.data.importing = true
+                            } else {
+                                instance.doImport(5)
+                            }
                         } else if( instance.models.skipMapping && !instance.models.skipImageSource ) {
                             instance.getImageFolders()
                             instance.stepTo(4)
@@ -402,8 +433,17 @@
                             this.data.map.headers = Object.keys(this.data.items[0])
                             this.data.loading = false
 
+                            if( 0 <= this.data.map.headers.indexOf('Image Link') ) {
+                                this.data.hasImageLinks = true
+                            }
+
                             if( this.models.skipMapping && this.models.skipImageSource ) {
-                                this.doImport(5)
+                                if( !this.data.hasImageLinks ) {
+                                    this.doImport(5, true)
+                                    this.data.importing = true
+                                } else {
+                                    this.doImport(5)
+                                }
                             } else if( this.models.skipMapping && !this.models.skipImageSource ) {
                                 this.getImageFolders()
                                 this.stepTo(4)
@@ -414,11 +454,15 @@
                     })
                 }
             },
-            doImport(step = false) {
+            doImport(step = false, proceed = false) {
                 if( step ) {
                     this.stepTo(step)
                 }
                 
+                if( !proceed ) 
+                    return false
+
+                this.data.importing = true
                 this.data.loading = true
 
                 axios({
@@ -430,6 +474,7 @@
                             nonce           : mosWC.ajax.nonce,
                             items           : this.data.items.slice((this.data.imported.batch - 1), this.data.imported.batch),
                             map             : this.data.map,
+                            importProducts  : this.data.importProducts ? 1 : 0,
                             imageSources    : 0 < this.data.sources.length ? this.data.sources : '',
                             time            : this.data.time
                         }
@@ -440,15 +485,21 @@
                     if( response.data.skipped || 
                         this.data.imported.items.length < this.data.items.length ) {
                         this.data.imported.batch += 1
-                        this.doImport(step)
+                        this.doImport(step, true)
                     } else {
-                        this.getUnimported()
+                        if( this.data.importProducts ) {
+                            this.getUnimported(true)
+                        } else {
+                            this.data.loading = false
+
+                            this.stepTo(6)
+                        }
                     }
                 }).catch(error => {
                     console.log(error)
                 })
             },
-            getUnimported() {
+            getUnimported(doDelete = false) {
                 axios({
                     url     : mosWC.ajax.url,
                     method  : 'POST',
@@ -462,7 +513,14 @@
                     if( 0 < response.data.items.length ) {
                         this.data.unimported.items = this.data.unimported.items.concat(response.data.items)
 
-                        this.deleteUnimported(this.data.unimported.items)
+                        if( doDelete ) {
+                            this.deleteUnimported(this.data.unimported.items)
+                        }
+                    } else {
+                        this.data.loading = false
+                        this.data.cleaningUp = false
+
+                        this.stepTo(6)
                     }
                 }).catch(error => {
                     console.log(error)
